Add Room database

+++ app/src/main/java/com/redcatgames/musiclib/domain/repository/ArtistRepository.kt
@@ -4,5 +4,5 @@ import com.redcatgames.musiclib.domain.model.Artist
 
 interface ArtistRepository {
     fun getArtistList(): List<Artist>
-    fun getArtist(id: Int): Artist
+    fun getArtist(id: Long): Artist
 }
