+++ app/src/main/java/com/redcatgames/musiclib/domain/usecase/GetArtistUseCase.kt
@@ -4,5 +4,5 @@ import com.redcatgames.musiclib.domain.repository.ArtistRepository
 import javax.inject.Inject
 
 class GetArtistUseCase @Inject constructor(private val artistRepository: ArtistRepository) {
-    operator fun invoke(id: Int) = artistRepository.getArtist(id)
+    operator fun invoke(id: Long) = artistRepository.getArtist(id)
 }
